update, refactory response api in category module

+++ src/category/category.service.ts
@@ -1,6 +1,6 @@
 import { PrismaService } from 'src/prisma/prisma.service';
 import { Category } from './category.model';
-import { Inject } from '@nestjs/common';
+import { Inject, HttpStatus, HttpException } from '@nestjs/common';
 
 export class CategoryService {
   constructor(
@@ -8,36 +8,80 @@ export class CategoryService {
     private prisma: PrismaService,
   ) {}
 
-  async getAllCategory(): Promise<Category[]> {
-    return this.prisma.category.findMany();
+  async getAllCategory() {
+    const data = await this.prisma.category.findMany();
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Success',
+      data: data,
+    };
   }
 
-  async getCategoryById(id: number): Promise<Category | null> {
-    return this.prisma.category.findFirst({ where: { id: Number(id) } });
+  async getCategoryById(id: number) {
+    const data = await this.prisma.category.findFirst({
+      where: { id: Number(id) },
+    });
+
+    if (!data) {
+      throw new HttpException('Data not found', HttpStatus.NOT_FOUND);
+    }
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Success',
+      data: data,
+    };
   }
 
-  async createCategory(data: Category): Promise<Category> {
-    return this.prisma.category.create({ data });
+  async createCategory(payload: Category) {
+    const data = await this.prisma.category.create({ data: payload });
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Success',
+      data: data,
+    };
   }
 
-  async updateCategory(id: number, data: Category): Promise<Category> {
-    const findData = this.prisma.category.findFirst({
+  async updateCategory(id: number, payload: Category) {
+    const findData = await this.prisma.category.findFirst({
       where: { id: Number(id) },
     });
 
     if (!findData) {
-      return null;
+      throw new HttpException('Data not found', HttpStatus.NOT_FOUND);
     }
 
-    return this.prisma.category.update({
+    const data = await this.prisma.category.update({
       where: { id: Number(id) },
-      data: { name: data.name },
+      data: { name: payload.name },
     });
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Success',
+      data: data,
+    };
   }
 
-  async deleteCategory(id: number): Promise<Category> {
-    return this.prisma.category.delete({
+  async deleteCategory(id: number) {
+    const findData = await this.prisma.category.findFirst({
       where: { id: Number(id) },
     });
+
+    if (!findData) {
+      throw new HttpException('Data not found', HttpStatus.NOT_FOUND);
+    }
+
+    await this.prisma.category.delete({
+      where: { id: Number(id) },
+    });
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Success',
+      data: id,
+    };
   }
 }
